Refactor and move all endpoints, remove sync APIs
Broken up all api endpoints across a few modules
in the api directory. Removed all synchronous
implementations of api endpoints for now.
